Remove misleading statement
The cited blog post according to which publishers would be harmed at n% by the removal of cookies is plainly and simply bad science. The way in which one can evaluate the value of cookie tracking is by comparing a world in which it is available to a world in which it isn't. What the linked blog does is measure the value of cookies in a world in which people can bid on them elsewhere. The findings it trumpets is not surprising to anyone, it is also completely unrelated to the statement being made. It tells us nothing about what the impact on publishers would be if Chrome were to implement ITP. On the contrary, there are many good reasons to expect that it would in fact be beneficial.